EnemyMeleeAttack
a melee box appears and pushs away player and damages player. The box appears based off of range to the player and the monster now rotates towards player

P.S.
I honeslty dont know if this is what u imagined it to be. please check it out and see how its implenmented and if it fits what u guys imagine