Map grid config and spawning
THe Map is now configured to a 6x6 grid so that the system can track the player movement. Currently the uses a 2d int array to represent its postion and the int array contains the number type of the room 1-4. There is a visual in the script file roomsetup under the door folder. only room type one is active and useable atm.